# backend/src/main/resources/openapi.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/transactions/{id}": {
      "put": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "updateTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "deleteTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/transactions": {
      "post": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "createTransaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/spaces": {
      "get": {
        "tags": [
          "space-controller"
        ],
        "operationId": "getSpaces",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpaceResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "space-controller"
        ],
        "operationId": "createSpace",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/spaces/{spaceId}": {
      "patch": {
        "tags": [
          "space-controller"
        ],
        "operationId": "updateSpaceName",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "newName",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/transactions/search": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "searchTransactions",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "notes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "needReview",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "dateFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "dateTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseTransactionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/spaces/switch/{spaceId}": {
      "get": {
        "tags": [
          "space-controller"
        ],
        "operationId": "switchSpace",
        "parameters": [
          {
            "name": "spaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SpaceResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "UpdateTransactionRequest": {
        "required": [
          "date",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME",
              "TRANSFER"
            ]
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "fromAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "toAccountId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AccountResponse": {
        "required": [
          "balance",
          "currency",
          "isBankAccount",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "balance": {
            "type": "number"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyResponse"
          },
          "type": {
            "$ref": "#/components/schemas/AccountTypeResponse"
          },
          "isBankAccount": {
            "type": "boolean"
          }
        }
      },
      "AccountTypeResponse": {
        "required": [
          "fullName",
          "iconName",
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          },
          "iconName": {
            "type": "string"
          }
        }
      },
      "CategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME"
            ]
          },
          "subCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponse"
            }
          }
        }
      },
      "CurrencyResponse": {
        "required": [
          "code",
          "fullName",
          "shortName",
          "sign"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "fullName": {
            "type": "string"
          },
          "shortName": {
            "type": "string"
          },
          "sign": {
            "type": "string"
          }
        }
      },
      "TransactionResponse": {
        "required": [
          "account",
          "amount",
          "currency",
          "date",
          "id",
          "isBankTransaction",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "$ref": "#/components/schemas/CurrencyResponse"
          },
          "account": {
            "$ref": "#/components/schemas/AccountResponse"
          },
          "fromAccount": {
            "$ref": "#/components/schemas/AccountResponse"
          },
          "toAccount": {
            "$ref": "#/components/schemas/AccountResponse"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryResponse"
          },
          "type": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME",
              "TRANSFER"
            ]
          },
          "isBankTransaction": {
            "type": "boolean"
          }
        }
      },
...
      "CreateTransactionRequest": {
        "required": [
          "accountId",
          "amount",
          "currency",
          "date",
          "name",
          "type"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "enum": [
              "UAH",
              "USD",
              "EUR",
              "TRY"
            ]
          },
          "amount": {
            "type": "number"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "fromAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "toAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME",
              "TRANSFER"
            ]
          }
        }
      },
      "SpaceRequest": {
        "required": [
          "name",
          "primaryCurrencyCode"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "primaryCurrencyCode": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SpaceResponse": {
        "required": [
          "id",
          "name",
          "primaryCurrency"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "primaryCurrency": {
            "$ref": "#/components/schemas/CurrencyResponse"
          }
        }
      },
...
      "PageResponseTransactionResponse": {
        "required": [
          "content",
          "last",
          "pageNumber",
          "pageSize",
          "totalElements",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}